Crate flow_sdk

Flow.rs

This crate contains various ways to interact with the Flow blockchain through its access API.

Start by connecting to the testnet() or the mainnet(), or login to an account by providing the address and secret key.

You can also take a look at the reference for examples around individual functionalities.

Modules

access

Access API type definitions

account

Mutate the blockchain with your private key(s).

algorithms

The hashing and signing algorithms.

client

Flow gRPC connections

codec

tonic::codec::Codec implementation for otopr.

entities

Entities of the Flow network.

error

Errors

multi

Multi-party signing

prelude

Re-exports for commonly used types and functions.

protobuf

Protobuf objects to decode/encode.

requests

A trait to generalize requests and responses of the Access API.

sign

Common logic for signing transactions.

transaction

Transactions on the flow blockchain.