Module flow_sdk::access

Access API type definitions

This modules contains definitions for requests and responses of the access API.

Structs

AccountResponse

An account.

BlockHeaderResponse

A block header.

BlockResponse

Full information about a block.

CollectionResponse

A collection.

EventsResponse

Search results for events in multiple blocks.

EventsResult

Search results for events in a single block.

ExecuteScriptAtBlockHeightRequest

Executes a script (maybe with arguments) at a block specified by its height.

ExecuteScriptAtBlockIdRequest

Executes a script (maybe with arguments) at a block specified by its block ID.

ExecuteScriptAtLatestBlockRequest

Executes a script (maybe with arguments) at the latest block.

ExecuteScriptResponse

The return value of the script.

ExecutionResultForBlockIdResponse

An execution result.

GetAccountAtBlockHeightRequest

Retrieves information of an account at the specific block height.

GetAccountAtLatestBlockRequest

Retrieves information of an account at the latest block.

GetBlockByHeightRequest

Gets full information about a block by its height.

GetBlockByIdRequest

Gets full information about a block by its id.

GetBlockHeaderByHeightRequest

Gets a block’s header by its height.

GetBlockHeaderByIdRequest

Gets a block’s header by id.

GetCollectionByIdRequest

Gets information about a collection by its id.

GetEventsForBlockIdsRequest

Search for events in a collection of blocks specified by its block id.

GetEventsForHeightRangeRequest

Search for events in a height range.

GetExecutionResultForBlockIdRequest

Retrieves execution result for given block. It is different from Transaction Results, and contain data about chunks/collection level execution results rather than particular transactions. Particularly, it contains EventsCollection hash for every chunk which can be used to verify the events for a block.

GetLatestBlockHeaderRequest

Gets the latest block’s header.

GetLatestBlockRequest

Gets full information about the latest block.

GetLatestProtocolStateSnapshotRequest

Retrieves the latest Protocol state snapshot serialized as a byte array.

GetNetworkParametersRequest

Get network parameters.

GetNetworkParametersResponse

The network parameters.

GetTransactionRequest

Gets a transaction’s details by its id.

PingRequest

Ping.

PingResponse

Pong.

ProtocolStateSnapshotResponse

A protocol state snapshot.

SendTransactionRequest

Sends a transaction over the network.

SendTransactionResponse

The id of the transaction on the network.

TransactionResponse

The full details of a transaction.

TransactionResultResponse

The results of a transaction.