Module flow_sdk::algorithms

The hashing and signing algorithms.

Modules

rand

Re-exports items from the rand crate.

secp256k1

Re-exports items from the secp256k1 crate.

sha3

Re-exports items from the tiny_keccak crate.

Structs

P256

P256 / secp256r1 curve.

Secp256k1

secp256k1 curve.

Sha2

SHA2 256 bit hashing.

Sha3

SHA3 256 bit hashing.

Traits

FlowHasher

A hasher.

FlowSigner

A signature signer.

HashAlgorithm

A hashing algorithm.

SecretKey

A secret key.

Signature

A signature.

SignatureAlgorithm

A signature algorithm.

Type Definitions

DefaultHasher

The default hasher, the exact type depends on the feature flags enabled.

DefaultSecretKey

The default secret key, the exact type depends on the feature flags enabled.

DefaultSigner

The default signer, the exact type depends on the feature flags enabled.