Module flow_sdk::client

Flow gRPC connections

This module contains the Client types for gRPC connections.

If you wish to customize and build your own client, implement GrpcClient for your client for input and output types you want to support. If you can support all types, consider using the FlowRequest trait to generalize implementations.

Structs

FlowClient

A gRPC client wrapper. Has utility functions for sending requests.

Traits

GrpcClient

A gRPC client trait.

Type Definitions

GrpcSendResult

The return type of sending a request over the gRPC connection.

TonicClient

A client that uses the tonic gRPC dispatcher which wraps some inner gRPC service.

TonicFlowClient

A tonic gRPC client.

TonicHyperClient

A tonic gRPC client that uses the hyper crate for HTTP transport.

TonicHyperFlowClient

A flow client that uses TonicHyperClient as gRPC client.