Trait flow_sdk::client::GrpcClient

pub trait GrpcClient<I, O> {
    type Error: Into<Box<dyn Error + Send + Sync>>;
    fn send<'a>(
        &'a mut self, 
        input: I
    ) -> Pin<Box<dyn Future<Output = Result<O, Self::Error>> + 'a>>;
}

A gRPC client trait.

Implementors should be generic over the input and output types, but it is not required.

Associated Types

type Error: Into<Box<dyn Error + Send + Sync>>

The error type of the client.

Required methods

fn send<'a>(
&'a mut self,
input: I
) -> Pin<Box<dyn Future<Output = Result<O, Self::Error>> + 'a>>

Sends a request with the client. Returns a future that evaluates a Result, potentially containing the output.