Module flow_sdk::entities

Entities of the Flow network.

This module contains data entities returned or accepted by the Access API.

Structs

Account

An account is a user’s identity on Flow. It contains a unique address, a balance, a list of public keys and the code that has been deployed to the account.

AccountKey

A key configured on some account.

Block

A block.

BlockHeader

A block header is a summary of a block and contains only the block ID, height, and parent block ID.

BlockSeal

A block seal is an attestation that the execution result of a specific Block has been verified and approved by a quorum of verification nodes.

Chunk

Chunk describes execution information for given collection in a block.

Collection

A collection is a batch of transactions that have been included in a block.

CollectionGuarantee

A collection guarantee is a signed attestation that specifies the collection nodes that have guaranteed to store and respond to queries about a collection.

Event

An event is emitted as the result of a transaction execution.

ExecutionResult

Execution result for a particular block.

ServiceEvent

Special type of events emitted in system chunk used for controlling Flow system.