Struct flow_sdk::multi::PartyBuilder

pub struct PartyBuilder { /* fields omitted */ }

A builder that makes it easy to create new SigningParty instances.

let party = PartyBuilder::new()
    .script("s")
    .reference_block([0])
    .gas_limit(123)
    .proposer_address([1])
    .proposal_key_id(2)
    .proposal_key_sequence_number(3)
    .payer([4])
    .authorizer([5])
    .build();

let party2 = SigningParty::new("s".into(), [].into(), [0].into(), 123, [1].into(), 2, 3, [4].into(), [[5].into()].into());

assert_eq!(party, party2);

Implementations

impl PartyBuilder

pub const fn new() -> Self

Creates a new PartyBuilder with the default gas limit of 1000.

pub fn script(self, script: impl Into<Box<str>>) -> Self

Sets the script of the transaction.

pub fn argument(self, argument: impl Serialize) -> Self

Appends a new argument.

pub fn arguments(
self,
arguments: impl IntoIterator<Item = impl Serialize>
) -> Self

Appends arguments.

pub fn argument_raw(self, argument: impl Into<Box<[u8]>>) -> Self

Appends a new UTF-8 encoded argument in Cadence JSON interchange format.

pub fn arguments_raw(
self,
arguments: impl IntoIterator<Item = impl Into<Box<[u8]>>>
) -> Self

Appends raw UTF-8 encoded arguments in Cadence JSON interchange format.

pub fn reference_block(self, reference_block: impl Into<Box<[u8]>>) -> Self

Sets the reference block for this transaction.

pub async fn latest_block_as_reference<C: GrpcClient<GetLatestBlockHeaderRequest, BlockHeaderResponse>>(
self,
client: &mut C
) -> Result<Self, C::Error>

Uses the latest block as the reference block for this transaction.

pub fn gas_limit(self, limit: u64) -> Self

Sets the gas limit of this transation.

pub fn proposer_address(self, addr: impl Into<Box<[u8]>>) -> Self

Sets the address of the account that proposes this transaction.

pub fn proposal_key_id(self, id: u64) -> Self

Sets the key id of the proposal key of this transaction.

pub fn proposal_key_sequence_number(self, sequence_number: u64) -> Self

Sets the sequence number of the proposal key of this transaction.

pub async fn proposer_account<'a, C, Sk, Sign, Hash>(
self,
acc: &'a mut Account<C, Sk, Sign, Hash>
) -> Result<Self, BoxError> where
C: GrpcClient<GetAccountAtLatestBlockRequest<'a>, AccountResponse>,
Sign: FlowSigner<SecretKey = Sk>,
Hash: FlowHasher,

Sets the address, key id and the sequence number by querying on the network about a logged-in account.

pub fn payer(self, address: impl Into<Box<[u8]>>) -> Self

Sets the address of the account that will pay for this transaction.

pub fn payer_account<C, S, Si, H>(self, acc: &Account<C, S, Si, H>) -> Self

Records the address of the account that will pay for this transaction.

pub fn authorizer(self, address: impl Into<Box<[u8]>>) -> Self

Appends the address of an account which authorizes this transaction.

pub fn authorizers(
self,
addresses: impl IntoIterator<Item = impl Into<Box<[u8]>>>
) -> Self

Appends the addresses of accounts that authorizes this transaction.

pub fn authorizer_account<C, S, Si, H>(self, acc: &Account<C, S, Si, H>) -> Self

Appends the address of an account which authorizers this transaction.

pub fn authorizer_accounts<'a, C: 'a, S: 'a, Si: 'a, H: 'a>(
self,
acc: impl IntoIterator<Item = &'a Account<C, S, Si, H>>
) -> Self

Appends the addresses of accounts that authorizes this transaction.

pub fn build(self) -> SigningParty

Builds a SigningParty from this builder, assuming all fields have been set.

pub fn build_prehashed<H: FlowHasher>(self) -> PreHashedParty<H>

Builds a PreHashedParty from this builder, assuming all fields have been set.