Struct flow_sdk::multi::SigningParty

pub struct SigningParty { /* fields omitted */ }

A basic signing party. Contains all the information needed to make signatures.

Implementations

impl SigningParty

pub fn build_signer_map(
proposer: &[u8],
payer: &[u8],
authorizers: &[impl AsRef<[u8]>]
) -> HashMap<Box<[u8]>, u32>

Builds a signer address to signer index map.

pub fn new(
script: Box<str>,
arguments: Box<[Box<[u8]>]>,
reference_block: Box<[u8]>,
gas_limit: u64,
proposer_address: Box<[u8]>,
proposal_key_id: u64,
proposal_key_sequence_number: u64,
payer: Box<[u8]>,
authorizers: Box<[Box<[u8]>]>
) -> Self

Creates a new SigningParty with all the data associated to a transaction.

pub fn into_prehashed<H: FlowHasher>(self) -> PreHashedParty<H>

Computes the payload hash of this party, and turns this into a PreHashedParty.