Trait flow_sdk::multi::Party

pub trait Party<H: FlowHasher>: Sized + Sealed {

    fn script(&self) -> &str;

    fn arguments(&self) -> &[Box<[u8]>];

    fn reference_block(&self) -> &[u8];

    fn gas_limit(&self) -> u64;

    fn proposer_address(&self) -> &[u8];

    fn proposal_key_id(&self) -> u64;

    fn proposer_sequence_number(&self) -> u64;

    fn payer(&self) -> &[u8];

    fn authorizers(&self) -> &[Box<[u8]>];

    fn payload(&self) -> H;

    fn envelope(&self) -> H;

    fn add_payload_signature(
        &mut self, 
        signer_address: Box<[u8]>, 
        signer_id: u32, 
        signature: [u8; 64]
    );

    fn into_transaction_with_envelope_signatures<SigAddr, Sig>(
        self, 
        signatures: impl IntoIterator<Item = SignatureE<SigAddr, Sig>>
    ) -> PartyTransaction<SigAddr, Sig>;

}

The Party trait. You can get information about the transaction you are signing and sign it by accepting some type that implements this trait.

This is Sealed, which means no foreign types may implement it, and it is not Object safe, so no one can create bad behaving trait objects which can make this trait insecure.

Required methods

fn script(&self) -> &str

Gets the script of the transaction.

fn arguments(&self) -> &[Box<[u8]>]

Gets the arguments of the transaction.

fn reference_block(&self) -> &[u8]

Gets the reference block of the transaction.

fn gas_limit(&self) -> u64

Gets the gas limit of the transaction.

fn proposer_address(&self) -> &[u8]

Gets the address of the proposer of the transaction.

fn proposal_key_id(&self) -> u64

Gets the key ID number of the proposal key.

fn proposer_sequence_number(&self) -> u64

Gets the sequence number of the proposal key.

fn payer(&self) -> &[u8]

Gets the address of the payer of the transaction.

fn authorizers(&self) -> &[Box<[u8]>]

Gets the addresses of the authorizers of the transaciton.

fn payload(&self) -> H

Computes the hash of the payload.

fn envelope(&self) -> H

Computes the hash of the envelope.

fn add_payload_signature(
&mut self,
signer_address: Box<[u8]>,
signer_id: u32,
signature: [u8; 64]
)

Adds a payload signature.

fn into_transaction_with_envelope_signatures<SigAddr, Sig>(
self,
signatures: impl IntoIterator<Item = SignatureE<SigAddr, Sig>>
) -> PartyTransaction<SigAddr, Sig>

Creates a PartyTransaction by feeding this party with envelope signatures.