Module flow_sdk::transaction

Transactions on the flow blockchain.

A transaction represents a unit of computation that is submitted to the Flow network.

Modules

rlp

RLP encoding of transactions.

Structs

AddContractTransaction

Adds a contract to an account.

CreateAccountTransaction

A simple transaction to create an account with full weight keys.

CreateAccountWeightedTransaction

A transaction to create an account with weighed keys.

Finalize

Repeatedly queries a client about a transaction, yielding the result after it has been sealed or expired.

ParseArguments

An iterator that parses arguments from bytes.

ProposalKeyD

The proposal key is used to specify a sequence number for the transaction.

ProposalKeyE

The proposal key is used to specify a sequence number for the transaction.

RemoveContractTransaction

Remove a contract of an account by name.

SignatureD

Signature of a transaction.

SignatureE

Signature of a transaction.

TransactionD

A transaction represents a unit of computation that is submitted to the Flow network.

TransactionE

A transaction represents a unit of computation that is submitted to the Flow network.

TransactionHeader

A TransactionHeader is a template plus arguments.

TransactionHeaderBuilder

A builder for TransactionHeaders.

UpdateContractTransaction

Updates a contract of an account by name.

Enums

TransactionStatus

Status of a transaction.