Struct flow_sdk::transaction::TransactionD

pub struct TransactionD {
    pub script: Box<[u8]>,
    pub arguments: Repeated<Vec<Box<[u8]>>>,
    pub reference_block_id: Box<[u8]>,
    pub gas_limit: u64,
    pub proposal_key: ProposalKeyD,
    pub payer: Box<[u8]>,
    pub authorizers: Repeated<Vec<Box<[u8]>>>,
    pub payload_signatures: Repeated<Vec<SignatureD>>,
    pub envelope_signatures: Repeated<Vec<SignatureD>>,
}

A transaction represents a unit of computation that is submitted to the Flow network.

This type is used when decoding messages from the network.

Fields
script: Box<[u8]>

Raw source code for a Cadence script, encoded as UTF-8 bytes

arguments: Repeated<Vec<Box<[u8]>>>

Arguments passed to the Cadence script, encoded as JSON-Cadence bytes

reference_block_id: Box<[u8]>

Block ID used to determine transaction expiry

gas_limit: u64

The gas limit.

proposal_key: ProposalKeyD

Account key used to propose the transaction

payer: Box<[u8]>

Address of the payer account

authorizers: Repeated<Vec<Box<[u8]>>>

Addresses of the transaction authorizers

payload_signatures: Repeated<Vec<SignatureD>>

Signatures from all payload signer accounts

envelope_signatures: Repeated<Vec<SignatureD>>

Signatures from all envelope signer accounts

Implementations

impl TransactionD

pub fn parse_argument(&self, index: usize) -> Result<ValueOwned>

Parse a specific argument.

pub fn parse_arguments(&self) -> ParseArguments<Iter<'_, Box<[u8]>>>

Returns an iterator parsing the underlying arguments.